refactor python build system to deduplicate
