--- calico/__init__.py
@@ -1,2 +1,0 @@
-from ._calico import *
-from .utils import *
